Beginning to split things into separate jars
