+++ build.gradle
@@ -46,25 +46,21 @@ task wrapper(type: Wrapper) {
 
 project(':ds3-sdk') {
     dependencies {
-        compile project(':ds3-interfaces')
         compile project(':ds3-utils')
     }
 }
 
 project(':ds3-metadata') {
     dependencies {
-        compile project(':ds3-interfaces')
         compile project(':ds3-utils')
     }
 }
 
-/*
 project(':ds3-sdk-integration') {
     dependencies {
         compile project(':ds3-sdk')
     }
 }
-*/
 
 project(':ds3-sdk-samples') {
     dependencies {
