+++ ds3-metadata/build.gradle
@@ -1,14 +1,23 @@
+/*
+ * ******************************************************************************
+ *   Copyright 2014-2016 Spectra Logic Corporation. All Rights Reserved.
+ *   Licensed under the Apache License, Version 2.0 (the "License"). You may not use
+ *   this file except in compliance with the License. A copy of the License is located at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   or in the "license" file accompanying this file.
+ *   This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+ *   CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ *   specific language governing permissions and limitations under the License.
+ * ****************************************************************************
+ */
+
 dependencies {
-    compile     'org.apache.httpcomponents:httpclient:4.5.1'
-    compile     'commons-codec:commons-codec:1.10'
-    compile     'commons-io:commons-io:2.4'
-    compile     'org.codehaus.woodstox:woodstox-core-asl:4.4.1'
-    compile     'com.fasterxml.jackson.dataformat:jackson-dataformat-xml:2.7.1'
-    compile     'com.google.guava:guava:19.0'
-    compile     'com.google.code.findbugs:annotations:3.0.1'
-    compile group: 'net.java.dev.jna', name: 'jna-platform', version: '4.2.2'
-    compile group: 'net.java.dev.jna', name: 'jna', version: '4.2.2'
-    testCompile 'org.hamcrest:hamcrest-library:1.3'
     testCompile group: 'org.apache.commons', name: 'commons-lang3', version: '3.0'
+    compile     group: 'net.java.dev.jna', name: 'jna-platform', version: '4.2.2'
+    compile     group: 'net.java.dev.jna', name: 'jna', version: '4.2.2'
+    compile     'commons-io:commons-io:2.4'
+    compile     'org.apache.httpcomponents:httpclient:4.5.1'
 }
