+++ ds3-sdk/build.gradle
@@ -74,14 +74,10 @@ jar.dependsOn genConfigProperties
 
 dependencies {
     compile     'org.apache.httpcomponents:httpclient:4.5.1'
-    compile     'commons-codec:commons-codec:1.10'
     compile     'commons-io:commons-io:2.4'
-    compile     'org.codehaus.woodstox:woodstox-core-asl:4.4.1'
     compile     'com.fasterxml.jackson.dataformat:jackson-dataformat-xml:2.7.1'
     compile     'com.google.guava:guava:19.0'
     compile     'com.google.code.findbugs:annotations:3.0.1'
-    compile group: 'net.java.dev.jna', name: 'jna-platform', version: '4.2.2'
-    compile group: 'net.java.dev.jna', name: 'jna', version: '4.2.2'
     testCompile 'org.hamcrest:hamcrest-library:1.3'
     testCompile group: 'org.apache.commons', name: 'commons-lang3', version: '3.0'
 }
