+++ ds3-sdk/src/main/java/com/spectralogic/ds3client/helpers/strategy/StrategyUtils.java
@@ -24,9 +24,6 @@ import com.spectralogic.ds3client.models.Objects;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.List;
 import java.util.UUID;
