+++ ds3-utils/src/main/java/com/spectralogic/ds3client/utils/Platform.java
